Merge

+++ s3_servicioshogar-back/src/main/java/co/edu/uniandes/csw/servicioshogar/entities/ReferenciaEntity.java
@@ -18,8 +18,8 @@ public class ReferenciaEntity extends BaseEntity implements Serializable{
           
     private String empresa;
     private String nombreRemitente;
-    private int idRemitente;
-    private int telRemitente;
+    private Long idRemitente;
+    private Long telRemitente;
     private String cargo;
     private String email;
     private String parentesco;
@@ -41,19 +41,19 @@ public class ReferenciaEntity extends BaseEntity implements Serializable{
         this.nombreRemitente = nombreRemitente;
     }
 
-        public int getIdRemitente(){
+        public Long getIdRemitente(){
         return idRemitente;
     }
 
-    public void setIdRemitente(int idRemitente){
+    public void setIdRemitente(Long idRemitente){
         this.idRemitente = idRemitente;
     }
     
-    public int getTelRemitente() {
+    public Long getTelRemitente() {
         return telRemitente;
     }
 
-    public void setTelRemitente(int telRemitente) {
+    public void setTelRemitente(Long telRemitente) {
         this.telRemitente = telRemitente;
     }
